[WKCI][browserperfdash-benchmark] Pass the build log URL from buildbot to the runner and forward it to the dashboard https://bugs.webkit.org/show_bug.cgi?id=303542
Reviewed by Nikolas Zimmermann.

Pass the URL of the build log from buildbot to the browserperfdash-benchmark runner,
and forward it as part of the data that the runner send to the performance dashboard.

The idea is that we will use this URL data to show it on the datapoints, so its easier
to open the buildbot logs that generated a given data point on the performance dashboard.

As part of this modernize a bit the builbot RunBenchmarkTests step and fix some issues.
The previous getText2() method was not working to generate the result summary string in
case of failure. Also passing the name of the browser is not longer needed as now that
goes into the configuration file. Also add unit tests.

The timeout of the step is also rised to 2000 because the runners run with an internal
timeout per-test of 1984, so that way we ensure that the runner timeout triggers first.

* Tools/CISupport/build-webkit-org/steps.py:
(RunBenchmarkTests.__init__):
(RunBenchmarkTests.run):
(RunBenchmarkTests.evaluateCommand):
(RunBenchmarkTests.getResultSummary):
(RunBenchmarkTests.getText): Deleted.
(RunBenchmarkTests.getText2): Deleted.
* Tools/CISupport/build-webkit-org/steps_unittest.py:
* Tools/Scripts/webkitpy/browserperfdash/browserperfdash_runner.py:
(parse_args):
(BrowserPerfDashRunner.__init__):

Canonical link: https://commits.webkit.org/303904@main