[WKCI][build.webkit.org] Drop the custom step for running API tests on GTK and WPE and use the general one https://bugs.webkit.org/show_bug.cgi?id=303557
Reviewed by Aakash Jain.

GTK and WPE ports use a custom step for running the API tests on the config
of build.webkit.org, but not on the one of ews-build.webkit.org.

This patch drops the custom RunGLibAPITests() step and makes the GTK and WPE
ports use the generic RunAPITests() one instead.

Thanks to this the GTK and WPE runners will now generate the json with the
results, and the result summary string in case of failure should work now.
This was broken on the RunGLibAPITests() removed due to that step using
the legacy getText2() method instead of getResultSummary()

Also the custom high timeout on the step is removed, because since 303319@main
the tool filter-test-logs outputs each 5 minutes a summary of the number of
lines filtered, so it is not longer needed to rise the buildbot step timeouts
to avoid issues with the "timeout-without-output" timer.

* Tools/CISupport/build-webkit-org/factories.py:
(TestFactory.__init__):
(TestLayoutAndAPIOnlyFactory.__init__):
* Tools/CISupport/build-webkit-org/factories_unittest.py:
(TestExpectedBuildSteps):
* Tools/CISupport/build-webkit-org/steps.py:
(RunAPITests):
(RunAPITests.__init__):
(RunAPITests.run):
(RunMVTTests.getResultSummary):
(RunGLibAPITests): Deleted.
(RunGLibAPITests.run): Deleted.
(RunGLibAPITests.getText): Deleted.
(RunGLibAPITests.getText2): Deleted.
(RunGtkAPITests): Deleted.
(RunWPEAPITests): Deleted.
* Tools/CISupport/build-webkit-org/steps_unittest.py:

Canonical link: https://commits.webkit.org/303928@main