Layout test `pointer-tracking-disabled.html` scrolls the page with touch event regions enabled https://bugs.webkit.org/show_bug.cgi?id=303567 rdar://165857597
Reviewed by Tim Horton.

The purpose of the layout test is to ensure that the thumb of a disabled switch
cannot be dragged. However, the drag performed has an unnecessary vertical offset.
When touch event regions are enabled, touch event regions are not created for
disabled switch elements. This results in the page scrolling due to the drag.

Fix this issue by removing the vertical offset from the drag.

* LayoutTests/fast/forms/switch/pointer-tracking-disabled.html:

Canonical link: https://commits.webkit.org/303935@main