Improve HTMLTreeBuilder for table markup https://bugs.webkit.org/show_bug.cgi?id=303546 rdar://165838124
Reviewed by Ryosuke Niwa.

The current code had 3 FIXME pointing at the potential slowness of the
code. This improves and reduces slightly the performance isssue. It's
probably not a huge improvement. The A/B perf measurement is showing a
bit of improvement.

The code was going a couple through the same list of elements, instead
of doing one pass. This is fixing it.

* Source/WebCore/html/parser/HTMLElementStack.cpp:
(WebCore::HTMLElementStack::hasAnyInTableScope const):
* Source/WebCore/html/parser/HTMLElementStack.h:
* Source/WebCore/html/parser/HTMLTreeBuilder.cpp:
(WebCore::HTMLTreeBuilder::processStartTag):
(WebCore::HTMLTreeBuilder::processEndTagForInTableBody):

Canonical link: https://commits.webkit.org/303950@main