Switch to WeakPtr for WebPageProxy in TextCheckerCompletion. https://bugs.webkit.org/show_bug.cgi?id=303569 rdar://165860363
Reviewed by Wenson Hsieh.

When working on 303913@main one of the pieces of feedback
resulted in an empty class that then caused me to try and
use an empty Ref. This caused a crash. While the fix for that
issue was different, it exposed that the relationship between
TextCheckerCompletion and WebPageProxy was incorrect and we should
have an WeakPtr in the case that we've lost our WebPageProxy.

* Source/WebKit/UIProcess/TextCheckerCompletion.cpp:
(WebKit::TextCheckerCompletion::protectedPage const):
(WebKit::TextCheckerCompletion::spellDocumentTag):
(WebKit::TextCheckerCompletion::didFinishCheckingText const):
(WebKit::TextCheckerCompletion::didCancelCheckingText const):
* Source/WebKit/UIProcess/TextCheckerCompletion.h:

Canonical link: https://commits.webkit.org/303964@main