[CSS Zoom] line-height should scale font-relative units with text-zoom https://bugs.webkit.org/show_bug.cgi?id=303584 rdar://165073337
Reviewed by Sammy Gill.

As in https://commits.webkit.org/301757@main for the percentage case,
text zoom should scale the line-height when the property is set with
a font-relative unit when evaluationTimeZoomEnabled.

This is because, when evaluationTimeZoomEnabled is true,
computedSizeForRangeZoomOption returns the unzoomed font size.

Test: fast/text/line-height-em-text-zoom.html

* LayoutTests/fast/text/line-height-em-text-zoom-expected.html: Added.
* LayoutTests/fast/text/line-height-em-text-zoom.html: Added.
* Source/WebCore/style/values/inline/StyleLineHeight.cpp:
(WebCore::Style::CSSValueConversion<LineHeight>::operator):

Canonical link: https://commits.webkit.org/304012@main