[Site Isolation] WebContent processes in visible iframes are not always getting a foreground assertion https://bugs.webkit.org/show_bug.cgi?id=303660 rdar://165941185
Reviewed by Sihui Liu.

Local debugging shows that WebContent processes in visible iframes are not always getting a foreground assertion.
This is because a visible activity is not always taken on them. Visible activities are taken when the activity
state of the page changes, but this is not deterministic. This can be addressed by always taking a visible
activity on the WebContent process when creating a RemotePageProxy object.

* Source/WebKit/UIProcess/RemotePageProxy.cpp:
(WebKit::RemotePageProxy::injectPageIntoNewProcess):
* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::takeActivitiesOnRemotePage):
* Source/WebKit/UIProcess/WebPageProxy.h:

Canonical link: https://commits.webkit.org/304054@main