[Site Isolation] WebContent processes in iframes can start up without correct activity state https://bugs.webkit.org/show_bug.cgi?id=303692 rdar://165979299
Reviewed by Sihui Liu.

We aim to keep the activity states in sync between the main frame WebContent process and the iframe WebContent
processes. See e.g. WebPageProxy::takeVisibleActivity(), where the visible activity is being taken on the main
frame process and iframe processes. Currently, an iframe WebContent process can be started without the right
set of activities. In WebPageProxy::takeActivitiesOnRemotePage, we are only taking the visible and network
activity, but we should be taking the other activity types too if the main frame process has them. Otherwise,
we depend on the activities to be set when activity state is changing, which is not happening deterministically.

* Source/WebKit/UIProcess/WebPageProxy.cpp:
(WebKit::WebPageProxy::takeActivitiesOnRemotePage):

Canonical link: https://commits.webkit.org/304106@main