Make MessageEvent and ExtendableMessageEvent hold a SecurityOrigin https://bugs.webkit.org/show_bug.cgi?id=303629 rdar://165916784
Reviewed by Alex Christensen.

This is meant to be a strict refactoring driven by the needs of the
Origin API (bug 302785) which needs to be able to derive consistent
opaque origins from message events.

As such this should be covered by existing tests.

In MessageEvent.cpp we also tidy up an existing WTF::switchOn.

Canonical link: https://commits.webkit.org/304156@main