REGRESSION (301162@main): Focusing an invisible input element still draws the cursor with full opacity https://bugs.webkit.org/show_bug.cgi?id=302876 rdar://165489471
Reviewed by Aditya Keerthi.

After the changes in 301162@main, the text selection shows up even when focusing a hidden editable
element. Prior to that change, all codepaths that activated the selection would be gated on the text
selection interaction not being suppressed due to the editable root being hidden; however, we now
unconditionally activate the selection whenever the selection is reparented (when "selection honors
overflow scrolling" is enabled). To fix this, we add a check to only `-activateSelection` in the
case where the selection isn't being suppressed.

Already covered by 2 layout tests which are failing on internal bots (where
`SelectionHonorsOverflowScrolling` is enabled by default):

- editing/selection/ios/hide-selection-in-hidden-contenteditable-frame.html
- editing/selection/ios/hide-selection-in-hidden-contenteditable.html

* Source/WebKit/UIProcess/ios/WKContentViewInteraction.h:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
(-[WKContentView _isSuppressingSelectionAssistant]):

Expose text selection suppression state as a readonly getter, so that we can check it before
activating the selection below.

* Source/WebKit/UIProcess/ios/WKTextInteractionWrapper.mm:
(-[WKTextInteractionWrapper prepareToMoveSelectionContainer:]):

Canonical link: https://commits.webkit.org/304160@main