[GTK][WPE] Always use run loop observer to schedule rendering updates https://bugs.webkit.org/show_bug.cgi?id=303834
Reviewed by Nikolas Zimmermann.

Now that we use a run loop observer for the did composite notification,
we can just schedule the rendering update from didComposite, instead of
directly calling updateRendering(). This ensures update renderings are
always scheduled the same way. This optimization was important when we
used timers, because scheduling another timer from did composite meant
that the rendering update could happen after other previous scheudled
tasks or timers. With the run loop observer we know for sure that the
rendering update will be the next thing to be called, so the only
difference is that now we wake up the loop again.

Canonical link: https://commits.webkit.org/304174@main