[site-isolation] Fix document-leak-test.js to work with frames in different processes https://bugs.webkit.org/show_bug.cgi?id=303237
Reviewed by Ryosuke Niwa.

This adopts the fixes to http/tests/resources/document-leak-test.js from
r302683@main to the version in LayoutTests/resources, fixing with this
several tests that are failing when site isolation is enabled.

One important change from the version in http/tests/resources/ is that here
we run a proportional number of checks depending on how many frames
were loaded, since I noticed that a fixed number of checks are not always
going to be enough for GC to be triggered in time and are likely to cause
flakiness.

Also updated lazy-image-loading-document-leak.html to use this helper
instead of having its own implementation.

* LayoutTests/fast/dom/lazy-image-loading-document-leak-expected.txt:
* LayoutTests/fast/dom/lazy-image-loading-document-leak.html:
* LayoutTests/fast/dom/resources/lazy-image-loading-document-leak-frame.html: Renamed from LayoutTests/fast/dom/resources/lazy-image-loading-document-leak-popup.html.
* LayoutTests/platform/glib/media/media-source/remoteplayback-availability-callback-does-not-leak-expected.txt: Removed.
* LayoutTests/platform/ios-site-isolation/TestExpectations:
* LayoutTests/platform/mac-site-isolation/TestExpectations:
* LayoutTests/resources/document-leak-test.js:
(createFrames):
(iframeLeaked): Deleted.
(iframeSentMessage):
(runDocumentLeakTest): Deleted.

Canonical link: https://commits.webkit.org/304240@main