Frames which add touch/pointer/gesture event listeners to the window or document do not reliably generate touch event regions when touch event regions are enabled https://bugs.webkit.org/show_bug.cgi?id=303934 rdar://166072773
Reviewed by Tim Horton.

When a touch event listener is added to a document, we only invalidate the
event style for the parent document. We now invalidate the style for the document
for which the listener was actually added (and do the same when gesture events are
added).

* Source/WebCore/dom/Node.cpp:
(WebCore::tryAddEventListener):
(WebCore::didRemoveEventListenerOfType):
* Source/WebCore/page/LocalDOMWindow.cpp:
(WebCore::LocalDOMWindow::addEventListener):
(WebCore::LocalDOMWindow::removeEventListener):

Canonical link: https://commits.webkit.org/304257@main