Adopt isReleasedWeakValue in WeakPtr https://bugs.webkit.org/show_bug.cgi?id=303711 <rdar://problem/166015837>
Reviewed by Ryosuke Niwa.

This is a step toward a more efficient weak pointer.

The isReleasedWeakValue function makes HashTable responsible for shrinking the
table when it accumulates too many weak nulls.

ListHashSet was not compatible with this strategy. It depended on being notified
when an item was removed so that it could unlink and delete the item's linked
list node.

Changed ListHashSet to use a Linux-style sentinel / circular list. This has the
nice property that individual nodes can remove themselves from the list without
maintaining a back-pointer to the list owner.

Now the hash table owns the list nodes and destruction is automatic: Removing a
node deletes it, which removes it from the list.

Canonical link: https://commits.webkit.org/304265@main