Pass Connection pointer to Swift https://bugs.webkit.org/show_bug.cgi?id=303738 rdar://166048466
Reviewed by Richard Robinson.

The Swift IPC message handling code that was recently landed does not build
correctly for Swift message handlers taking a Connection parameter.
This change switches to a raw pointer to the Swift message handler function,
necessary for compatibility with the the generated function signature in the
generated header.

Canonical link: https://commits.webkit.org/304284@main