SpeculationRules - header only processed on main frame https://bugs.webkit.org/show_bug.cgi?id=303975
Reviewed by Alex Christensen.

According to the spec, the Speculation-Rules header should only be processed on the top-level frame.
This PR aligns the implementation to the spec, and gets the relevant test to pass.

No new tests, but progression on existing test.

* LayoutTests/imported/w3c/web-platform-tests/speculation-rules/header-only-top-level-expected.txt: Progression.
* Source/WebCore/loader/FrameLoader.cpp:
(WebCore::FrameLoader::receivedFirstData): Add an `isMainFrame()` condition.

Canonical link: https://commits.webkit.org/304309@main