[macOS] TestWebKitAPI.SiteIsolation.CancelOpenPanel is timing out in EWS rdar://163227871 https://bugs.webkit.org/show_bug.cgi?id=301313
Reviewed by Pascoe.

Sometimes the _doAfterProcessingAllPendingMouseEvents callback and the alert are received
close enough to each other that the TestWebKitAPI::Util::run call in TestWKWebView.waitForPendingMouseEvents
doesn't return between them, so we only start listening for the alert in the UI process after
the alert is received.  To fix this, don't wait for two things in a row, only wait for the alert.

* Tools/TestWebKitAPI/Tests/WebKitCocoa/SiteIsolation.mm:
(TestWebKitAPI::TEST(SiteIsolation, CancelOpenPanel)):

Canonical link: https://commits.webkit.org/304327@main